+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Button } from "./Button";
+export { default as Input } from "./Input";
+export { default as Login } from "./login";
+export { default as PwFind } from "./pwFind";
+export { default as SignUp } from "./signUp";
