+++ src/components/login.tsx
@@ -1,7 +1,6 @@
 import styled from "styled-components";
 import { useState } from "react";
-import Input from "../components/Input";
-import Button from "../components/Button";
+import { Input, Button } from "../components";
 import { useNavigate } from "react-router-dom";
 import { Arrow, LogoImg } from "../assets/index";
